+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # main
 
+* Added Cop `Migration/BulkChangeTable` 
+
 # v0.1.0
 
 * First version